add monolingual (en, spa) and multilingual 'not alphanum paragraph' taggers
